Rearchitect config file parsing
- Use a new custom parser class `SKLLConfigParser` based on `ConfigParser`.
- Move all of the configuration parsing code to `config.py` and out of `experiments.py`.
- Add a `validate()` method to the config parser that raises a KeyError if:
  - unrecognized options are specified
  - options are specified in more than one section
  - options are specified in the wrong section
- Add unit tests for the above validation checks
- Update test files to deal with the above changes
- Update `requirements_rtd.txt` to install the latest version of `ConfigParser` (v3.5.0b2)
