+++ tests/test_regression.py
@@ -27,12 +27,12 @@ from scipy.stats import pearsonr
 from sklearn.utils.testing import assert_greater, assert_less
 
 from skll.data import NDJWriter
-from skll.experiments import _setup_config_parser, run_configuration
+from skll.config import _setup_config_parser
+from skll.experiments import run_configuration
 from skll.learner import Learner
 from skll.learner import _DEFAULT_PARAM_GRIDS
 
-from utils import make_regression_data
-
+from utils import make_regression_data, fill_in_config_paths_for_fancy_output
 
 _ALL_MODELS = list(_DEFAULT_PARAM_GRIDS.keys())
 _my_dir = abspath(dirname(__file__))
@@ -404,35 +404,6 @@ def test_ensemble_models():
                use_rescaling)
 
 
-def fill_in_config_paths_for_fancy_output(config_template_path):
-    """
-    Add paths to train, test, and output directories to a given config template
-    file.
-    """
-
-    train_dir = join(_my_dir, 'train')
-    test_dir = join(_my_dir, 'test')
-    output_dir = join(_my_dir, 'output')
-
-    config = _setup_config_parser(config_template_path)
-
-    config.set("Input", "train_file", join(train_dir, "fancy_train.jsonlines"))
-    config.set("Input", "test_file", join(test_dir,
-                                              "fancy_test.jsonlines"))
-    config.set("Output", "results", output_dir)
-    config.set("Output", "log", output_dir)
-    config.set("Output", "predictions", output_dir)
-
-    config_prefix = re.search(r'^(.*)\.template\.cfg',
-                              config_template_path).groups()[0]
-    new_config_path = '{}.cfg'.format(config_prefix)
-
-    with open(new_config_path, 'w') as new_config_file:
-        config.write(new_config_file)
-
-    return new_config_path
-
-
 def test_int_labels():
     """
     Testing that SKLL can take integer input.
@@ -445,7 +416,7 @@ def test_int_labels():
     config_path = join(_my_dir, 'configs', 'test_int_labels_cv.cfg')
     output_dir = join(_my_dir, 'output')
 
-    config = _setup_config_parser(config_template_path)
+    config = _setup_config_parser(config_template_path, validate=False)
     config.set("Input", "train_file",
                join(_my_dir, 'other', 'test_int_labels_cv.jsonlines'))
     config.set("Output", "results", output_dir)
